int_8t is not 8 bits w/e and i'm stupid

+++ src/cpu/cpu.hpp
@@ -21,7 +21,7 @@ namespace Cpu {
         uint8_t D, E;
         uint8_t H, L;
         uint8_t S, P;
-        uint8_t PC;
+        uint16_t PC;
         uint8_t** const memory;
         std::unique_ptr<Parser> parser;
